source/requirements.txt
flask==3.1.1
python-dotenv==1.1.0
Flask-SQLAlchemy==3.1.1
mysqlclient==2.2.7
Flask-Migrate==4.1.0
httpx==0.28.1
pytest==8.4.0
pytest-dotenv==0.5.2